+++ src/PrestaShopBundle/Form/Admin/Sell/Order/Delivery/SlipOptionsType.php
@@ -23,41 +23,47 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  * International Registered Trademark & Property of PrestaShop SA
  */
+namespace PrestaShopBundle\Form\Admin\Sell\Order\Delivery;
 
-namespace PrestaShopBundle\Form\Admin\Order\Delivery;
-
-use PrestaShop\PrestaShop\Adapter\Order\Delivery\SlipPdfConfiguration;
-use PrestaShop\PrestaShop\Core\Form\FormDataProviderInterface;
+use PrestaShopBundle\Form\Admin\Type\SwitchType;
+use PrestaShopBundle\Form\Admin\Type\TranslateTextType;
+use PrestaShopBundle\Form\Admin\Type\TranslatorAwareType;
+use Symfony\Component\Form\Extension\Core\Type as FormType;
+use Symfony\Component\Form\FormBuilderInterface;
 
 /**
- * This class is responsible of managing the data manipulated pdf form
- * in "Sells > Orders > Delivery Slips" page.
+ * This form class generates the "Options" form in Delivery slips page
  */
-final class SlipPdfFormDataProvider implements FormDataProviderInterface
+class SlipOptionsType extends TranslatorAwareType
 {
     /**
-     * @var SlipPdfConfiguration
+     * {@inheritdoc}
      */
-    private $configuration;
-
-    public function __construct(SlipPdfConfiguration $configuration)
+    public function buildForm(FormBuilderInterface $builder, array $options)
     {
-        $this->configuration = $configuration;
+        $builder
+            ->add(
+                'prefix',
+                TranslateTextType::class,
+                [
+                    'locales' => $this->locales,
+                ]
+            )
+            ->add(
+                'number',
+                FormType\NumberType::class
+            )
+            ->add(
+                'enable_product_image',
+                SwitchType::class
+            );
     }
 
     /**
      * {@inheritdoc}
      */
-    public function getData()
+    public function getBlockPrefix()
     {
-        return [];
-    }
-
-    /**
-     * {@inheritdoc}
-     */
-    public function setData(array $data)
-    {
-        return $this->configuration->updateConfiguration($data['pdf']);
+        return 'order_delivery_slip_options';
     }
 }
