+++ src/PrestaShopBundle/Form/Admin/Sell/Order/Delivery/SlipPdfType.php
@@ -23,7 +23,7 @@
  * @license   https://opensource.org/licenses/OSL-3.0 Open Software License (OSL 3.0)
  * International Registered Trademark & Property of PrestaShop SA
  */
-namespace PrestaShopBundle\Form\Admin\Order\Delivery;
+namespace PrestaShopBundle\Form\Admin\Sell\Order\Delivery;
 
 use DateTime;
 use PrestaShopBundle\Form\Admin\Type\DatePickerType;
