+++ src/PrestaShopBundle/Resources/config/routing/admin/routing_order.yml
@@ -2,12 +2,12 @@ admin_order_delivery_slip:
     path: /delivery/slip
     methods:  [GET, POST]
     defaults:
-        _controller: PrestaShopBundle:Admin/Order/Delivery:slip
+        _controller: PrestaShopBundle:Admin/Sell/Order/Delivery:slip
         _legacy_controller: AdminDeliverySlip
 
 admin_order_delivery_slip_pdf:
     path: /delivery/slip/pdf
     methods:  [POST]
     defaults:
-        _controller: PrestaShopBundle:Admin/Order/Delivery:generatePdf
+        _controller: PrestaShopBundle:Admin/Sell/Order/Delivery:generatePdf
         _legacy_controller: AdminDeliverySlip
