+++ src/PrestaShopBundle/Resources/config/services/bundle/form/form_data_provider.yml
@@ -64,11 +64,11 @@ services:
             - '@prestashop.core.import.file_finder'
 
     prestashop.adapter.order.delivery.slip.options.form_provider:
-        class: 'PrestaShopBundle\Form\Admin\Order\Delivery\SlipOptionsFormDataProvider'
+        class: 'PrestaShopBundle\Form\Admin\Sell\Order\Delivery\SlipOptionsFormDataProvider'
         arguments:
             - '@prestashop.adapter.order.delivery.slip.options.configuration'
 
     prestashop.adapter.order.delivery.slip.pdf.form_provider:
-        class: 'PrestaShopBundle\Form\Admin\Order\Delivery\SlipPdfFormDataProvider'
+        class: 'PrestaShopBundle\Form\Admin\Sell\Order\Delivery\SlipPdfFormDataProvider'
         arguments:
             - '@prestashop.adapter.order.delivery.slip.pdf.configuration'
